removes duplicate code in previous folder names
