--- basic_contract_caller/frontend/src/Global.css
@@ -1,7 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-body {
-    background-color: #1A1452;
-}
