--- basic_contract_caller/frontend/src/vite-env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="vite/client" />
