--- contract_transfer/frontend/README.md
@@ -1,4 +1,0 @@
-# Have Questions?
-
-For any questions about building front end applications with [useink](https://use.ink/frontend/overview/), join the [Element chat](https://matrix.to/#/%23useink:parity.io).
-
