--- contract_transfer/frontend/src/assets/contract_transfer.json
@@ -1,333 +1,0 @@
-{
-  "source": {
-    "hash": "0x3431764079ef7dacc018ed5bfee19e333caf631c92ab621e73d7ed21bed1bc93",
-    "language": "ink! 4.2.0",
-    "compiler": "rustc 1.70.0",
-    "build_info": {
-      "build_mode": "Release",
-      "cargo_contract_version": "2.2.1",
-      "rust_toolchain": "stable-x86_64-apple-darwin",
-      "wasm_opt_settings": {
-        "keep_debug_symbols": false,
-        "optimization_passes": "Z"
-      }
-    }
-  },
-  "contract": {
-    "name": "contract_transfer",
-    "version": "4.2.0",
-    "authors": [
-      "Parity Technologies <admin@parity.io>"
-    ]
-  },
-  "spec": {
-    "constructors": [
-      {
-        "args": [],
-        "default": false,
-        "docs": [
-          "Creates a new instance of this contract."
-        ],
-        "label": "new",
-        "payable": true,
-        "returnType": {
-          "displayName": [
-            "ink_primitives",
-            "ConstructorResult"
-          ],
-          "type": 0
-        },
-        "selector": "0x9bae9d5e"
-      }
-    ],
-    "docs": [],
-    "environment": {
-      "accountId": {
-        "displayName": [
-          "AccountId"
-        ],
-        "type": 4
-      },
-      "balance": {
-        "displayName": [
-          "Balance"
-        ],
-        "type": 3
-      },
-      "blockNumber": {
-        "displayName": [
-          "BlockNumber"
-        ],
-        "type": 9
-      },
-      "chainExtension": {
-        "displayName": [
-          "ChainExtension"
-        ],
-        "type": 10
-      },
-      "hash": {
-        "displayName": [
-          "Hash"
-        ],
-        "type": 7
-      },
-      "maxEventTopics": 4,
-      "timestamp": {
-        "displayName": [
-          "Timestamp"
-        ],
-        "type": 8
-      }
-    },
-    "events": [],
-    "lang_error": {
-      "displayName": [
-        "ink",
-        "LangError"
-      ],
-      "type": 2
-    },
-    "messages": [
-      {
-        "args": [
-          {
-            "label": "value",
-            "type": {
-              "displayName": [
-                "Balance"
-              ],
-              "type": 3
-            }
-          }
-        ],
-        "default": false,
-        "docs": [
-          " Transfers `value` amount of tokens to the caller.",
-          "",
-          " # Errors",
-          "",
-          " - Panics in case the requested transfer exceeds the contract balance.",
-          " - Panics in case the requested transfer would have brought this contract's",
-          "   balance below the minimum balance (i.e. the chain's existential deposit).",
-          " - Panics in case the transfer failed for another reason."
-        ],
-        "label": "give_me",
-        "mutates": true,
-        "payable": false,
-        "returnType": {
-          "displayName": [
-            "ink",
-            "MessageResult"
-          ],
-          "type": 0
-        },
-        "selector": "0x499bb739"
-      },
-      {
-        "args": [],
-        "default": false,
-        "docs": [
-          " Asserts that the token amount sent as payment with this call",
-          " is exactly `10`. This method will fail otherwise, and the",
-          " transaction would then be reverted.",
-          "",
-          " # Note",
-          "",
-          " The method needs to be annotated with `payable`; only then it is",
-          " allowed to receive value as part of the call."
-        ],
-        "label": "was_it_ten",
-        "mutates": false,
-        "payable": true,
-        "returnType": {
-          "displayName": [
-            "ink",
-            "MessageResult"
-          ],
-          "type": 0
-        },
-        "selector": "0xcafebabe"
-      }
-    ]
-  },
-  "storage": {
-    "root": {
-      "layout": {
-        "struct": {
-          "fields": [],
-          "name": "GiveMe"
-        }
-      },
-      "root_key": "0x00000000"
-    }
-  },
-  "types": [
-    {
-      "id": 0,
-      "type": {
-        "def": {
-          "variant": {
-            "variants": [
-              {
-                "fields": [
-                  {
-                    "type": 1
-                  }
-                ],
-                "index": 0,
-                "name": "Ok"
-              },
-              {
-                "fields": [
-                  {
-                    "type": 2
-                  }
-                ],
-                "index": 1,
-                "name": "Err"
-              }
-            ]
-          }
-        },
-        "params": [
-          {
-            "name": "T",
-            "type": 1
-          },
-          {
-            "name": "E",
-            "type": 2
-          }
-        ],
-        "path": [
-          "Result"
-        ]
-      }
-    },
-    {
-      "id": 1,
-      "type": {
-        "def": {
-          "tuple": []
-        }
-      }
-    },
-    {
-      "id": 2,
-      "type": {
-        "def": {
-          "variant": {
-            "variants": [
-              {
-                "index": 1,
-                "name": "CouldNotReadInput"
-              }
-            ]
-          }
-        },
-        "path": [
-          "ink_primitives",
-          "LangError"
-        ]
-      }
-    },
-    {
-      "id": 3,
-      "type": {
-        "def": {
-          "primitive": "u128"
-        }
-      }
-    },
-    {
-      "id": 4,
-      "type": {
-        "def": {
-          "composite": {
-            "fields": [
-              {
-                "type": 5,
-                "typeName": "[u8; 32]"
-              }
-            ]
-          }
-        },
-        "path": [
-          "ink_primitives",
-          "types",
-          "AccountId"
-        ]
-      }
-    },
-    {
-      "id": 5,
-      "type": {
-        "def": {
-          "array": {
-            "len": 32,
-            "type": 6
-          }
-        }
-      }
-    },
-    {
-      "id": 6,
-      "type": {
-        "def": {
-          "primitive": "u8"
-        }
-      }
-    },
-    {
-      "id": 7,
-      "type": {
-        "def": {
-          "composite": {
-            "fields": [
-              {
-                "type": 5,
-                "typeName": "[u8; 32]"
-              }
-            ]
-          }
-        },
-        "path": [
-          "ink_primitives",
-          "types",
-          "Hash"
-        ]
-      }
-    },
-    {
-      "id": 8,
-      "type": {
-        "def": {
-          "primitive": "u64"
-        }
-      }
-    },
-    {
-      "id": 9,
-      "type": {
-        "def": {
-          "primitive": "u32"
-        }
-      }
-    },
-    {
-      "id": 10,
-      "type": {
-        "def": {
-          "variant": {}
-        },
-        "path": [
-          "ink_env",
-          "types",
-          "NoChainExtension"
-        ]
-      }
-    }
-  ],
-  "version": "4"
-}
