--- contract_transfer/frontend/src/constants.ts
@@ -1,2 +1,0 @@
-export const CONTRACT_ROCOCO_ADDRESS =
-  '5E6iJxxnf2v8p1mfc9UV8ojBYAQbJQyd5H1gQwvpxESZYEja';
