--- contract_transfer/frontend/tailwind.config.js
@@ -1,2 +1,0 @@
-import config from '../../ui/tailwind.config';
-export default config;
