--- contract_transfer/frontend/vite.config.ts
@@ -1,7 +1,0 @@
-import react from '@vitejs/plugin-react';
-import { defineConfig } from 'vite';
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react()],
-});
